Add the CSR Website's Api Actions (#12) ~ Add the CSR Website's Api Actions for interfacing with Customer Support WatchDog reports and Player Reports with actions and visualization
~ Add the /Api/V1/Stats route for Summary.htm, it will return an array
of data, with element 0 being the titles, and every other element being
data that falls under that title.

~ Add the /Api/V1/Report route to get information on a CS Report,
returns the ID of the report info, the ID of the report, the PlayerID,
the amount of Reports that Player received, and how many times they set
off watchdog.

~ Add the /Api/V1/Player/History route for playerHistory.htm, it will
return an array of PlayerHistoryModels, which include the time stamp,
the Client's IPv4, the Client's UUID, the ID of the exploit, the
DataModel Hack flags that was sent with WatchDog, the DataModel Hack
Rules that were sent with WatchDog, and a ExploitPath for greater
information on the exploit that was report on the CSPanel.

~ Add the /Api/V1/Report/List route for csreportmgr.htm, it will return
an array of reports broken down into months, they will give info on the
report entity, such as ID, Created and Updated, then they will give
information on the report's status, such as it being active or not, it
will return the start record date and end record date, and the amount of
times watchdog and player reports were recorded for that term.

~ Add the /Api/V1/Report/Meta route, identical to Report/List.

Co-Authored-by: Alex B. <bkordan@mfdlabs.com>
Co-Authored-by: Jack Daniels <jrigger123@mfdlabs.com>